requirements.txt
pillow
numpy
opencv-python
matplotlib
keras
tensorflow